+++ khoros/objects/attachments.py
@@ -6,7 +6,7 @@
 :Example:           ``payload = format_attachment_payload(titles, file_paths)``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     03 May 2020
+:Modified Date:     07 May 2020
 """
 
 import os
@@ -25,7 +25,7 @@ def construct_multipart_payload(message_json, file_paths):
     :type message_json: dict
     :param file_paths: The full path(s) to one or more attachment (e.g. ``path/to/file1.pdf``)
     :type file_paths: str, tuple, list, set
-    :returns: The full payload for the multipart API call as a list of tuples
+    :returns: The full payload for the multipart API call as a dictionary
 
               .. note:: See the `Requests Documentation <https://rsa.im/2SvMsya>`_ for added context on the return data.
 
@@ -33,15 +33,10 @@ def construct_multipart_payload(message_json, file_paths):
              :py:exc:`khoros.errors.exceptions.DataMismatchError`
     """
     file_paths = core_utils.convert_string_to_tuple(file_paths)
-    attachments_json = format_attachment_payload(file_paths)
-    message_json['attachments'] = attachments_json
+    message_json['data']['attachments'] = format_attachment_payload(file_paths)
     files_payload = get_file_upload_info(file_paths)
-    # message_json = {'api.request': message_json}
-    full_payload = [('json', ('api.request', json.dumps(message_json), 'application/json'))]
-    full_payload.extend(files_payload)
-    # print(f"MESSAGE JSON:\n{message_json}\n\nFILES PAYLOAD:\n{files_payload}")    # TODO: Remove print debugging
-    # return message_json, files_payload
-    print(f"FULL PAYLOAD: {full_payload}")
+    full_payload = {'api.request': (None, json.dumps(message_json, default=str), 'application/json')}
+    full_payload.update(files_payload)
     return full_payload
 
 
@@ -99,36 +94,9 @@ def get_file_upload_info(file_paths):
     :returns: A dictionary with the file upload information for the API call
     """
     file_paths = core_utils.convert_string_to_tuple(file_paths)
-    print(f"FILE PATHS:\n{file_paths}")     # TODO: Remove print debugging
-    files = _format_multiple_files(file_paths)
+    files, count = {}, 1
+    for path in file_paths:
+        # TODO: Dynamically define the MIME type
+        files[f'attachment{count}'] = (f'{os.path.basename(path)}', open(path, 'rb'))
+        count += 1
     return files
-
-
-def _format_single_file(_file_path):
-    """This function formats a single file for use in an API call.
-
-    :param _file_path: The individual file path
-    :type _file_path: str
-    :returns: Dictionary with the properly formatted API payload
-    :raises: :py:exc:`FileNotFoundError`
-    """
-    _files = {
-        'file': {'attachment1', open(_file_path, 'rb')}
-    }
-    return _files
-
-
-def _format_multiple_files(_file_paths):
-    """THis function formats multiple files for use in an API call.
-
-    :param _file_paths: The list of file paths
-    :type _file_paths: list, tuple, set
-    :returns: List of 2-tuples that have been properly formatted as API payload
-    :raises: :py:exc:`FileNotFoundError`
-    """
-    _file_paths = core_utils.convert_string_to_tuple(_file_paths)
-    _files, _count = [], 1
-    for _path in _file_paths:
-        _files.append(('file', (f'attachment{_count}', open(_path, 'rb'))))
-        _count += 1
-    return _files